remove log, rename hook file
