+++ apps/plugin-browser/src/pages/popup/popup-contents/action-center/infer-entities-action/inference-requests.tsx
@@ -15,7 +15,7 @@ import {
   darkModeInputBackgroundColor,
   darkModePlaceholderColor,
 } from "../../../../shared/dark-mode-values";
-import { useLocalStorage } from "../../../../shared/use-storage-sync";
+import { useLocalStorage } from "../../../../shared/use-local-storage";
 import { InferenceRequest } from "./inference-request";
 
 const InferenceRequestContainer = ({
